+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - feature/PASSBOLT-2371: Force the charset of the cake_sessions table in utf8
 - feature/PASSBOLT-2325: As system administrator I shouldn't be able to execute passbolt CLI commands as root
 - feature/PASSBOLT-2397: As system administrator I should see in the healthcheck if app/tmp content and app/webroot/img/public content are writable
+- feature/PASSBOLT-1991: As system administrator I should see in the healthcheck if the server key can be used for encrypting/decrypting
 
 ## [1.6.4] - 2017-08-31
 ### Fixed